Including SubParam, getting blocker.
The problem is that when going from ChorMTT to ChorProc,
we change from two modes to only a single mode.

This means that we have to map local types to their corresponding
global types. But we don't know the required location if we don't
track it in the local mode.

This means that we ought to have multiple local locations. And thus
also multiple box modalities, one for each location.

Then id=>(@u;[]u) locally, and ([]u;@u)=>id globally.