WIP trying to convert the Var translation to ChorProc into new \piS.
The problem is that here I do need to interconvert between \pi and \piS and this
only works if the type is an @. Which is not given in the global translVar where
we might be wanting to project a global, non-@ type from under a box into the
current local process.

What we do know, is that in such a case, the projection happens into a single
process, not a list of them. This might help to still get the resVar cases...